% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{age_land}
\alias{age_land}
\title{Age landscape}
\usage{
age_land(LAND, landscape_ini, layer)
}
\arguments{
\item{LAND}{The name of the landscape being changed}

\item{landscape_ini}{The name of the original landscape replacing}

\item{layer}{The layer that is being affected on the landscape}
}
\description{
How does the landscape change on its own over time steps?
For now, simply reverts a specified layer back to its original values
In other words, e.g., crops are annual and regrow undamaged each year
}
