% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmse_apply_control.R
\name{append_UROM_output}
\alias{append_UROM_output}
\title{append_UROM_output}
\usage{
append_UROM_output(dat, costs, old_output)
}
\arguments{
\item{dat}{gmse_apply_ style list with output to summarise and append to previous gmse_apply_summary(output).}

\item{costs}{updated manager costs, as set by user in the preceding time step.}

\item{old_output}{existing gmse_apply_summary() output data to append to.}
}
\value{
Updated gmse_apply_summary() output structure. The last line of the new output will be blank (NA) apart from the new true and observed resource population sizes. The actions taken and costs set will be added to the previous line, as they represent what happened in the previous time step.
}
\description{
Appends output to an output table set up by gmse_apply_summary() but treats this as "part-finished line"; i.e. following a gmse_apply_UROM() run, so following user-resource-observation-manager run. Assumes last output line is blank apart from true and observed resource pop size. Costs for the last line will be set as input as set by user during the previous time step (costs). Actions will be set as those taken by users in the new sim step in dat, and a new line will be added  with new true and observed resource counts.
}
\details{
Appends output to an output table set up by gmse_apply_summary() but treats this as "part-finished line"; i.e. following a gmse_apply_UROM() run, so following user-resource-observation-manager run. Assumes last output line is blank apart from true and observed resource pop size. Costs for the last line will be set as input as set by user during the previous time step (costs). Actions will be set as those taken by users in the new sim step in dat, and a new line will be added  with new true and observed resource counts.
}
\examples{
\dontrun{
step1 <- gmse_apply(get_res = "Full", scaring = TRUE);

}
}
