% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/be_user.R
\name{be_hunter}
\alias{be_hunter}
\title{Become a hunter on the landscape}
\usage{
be_hunter(OBSERVATION, AGENT, RESOURCES, LAND, PARAS, view, times)
}
\arguments{
\item{OBSERVATION}{The observation array produced by the observation function within GMSE}

\item{AGENT}{The array of agents produced in the main gmse() function}

\item{RESOURCES}{The resources array produced by the resource function within GMSE}

\item{LAND}{The landscape array on which interactions between resources and agents occur}

\item{PARAS}{The vector of parameters that hold global and dynamic parameter values used by GMSE}

\item{view}{The distance within which agents are able to observe resources on the landscape}

\item{times}{The number of times that resources are observed in the observation model of GMSE}
}
\value{
the_land A cols by rows landscape with randomly distributed cell types
}
\description{
This function allows the user of the GMSE software to insert themselves as a hunter in the simulation, allowing them to cull some number of resources in a time step as observed by the agent whose ID is 2
}
\examples{
\dontrun{
HUNT_OUTCOME <- be_hunter(OBSERVATION_r, AGENTS, RESOURCES, LANDSCAPE_r, 
paras, agent_view, times_observe);
}
}
