% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{chapman_est}
\alias{chapman_est}
\title{Chapman estimator of mark-recapture}
\usage{
chapman_est(observation, paras)
}
\arguments{
\item{observation}{The array of resource observations from the observation model, used to estimate abundance of resources}

\item{paras}{The vector of parameters that hold global and dynamic parameter values used by GMSE}
}
\value{
The Chapman estimator (which is also performed GMSE in the manager function) returns a list that includes resource population size estimates along with 95% confidence intervals
}
\description{
Estimates population size using simulated mark-recapture data produced by the
observation model of GMSE.
}
\examples{
\dontrun{
analysis <- chapman_est(observation=obs_t, paras = paras);
}
}
