% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmse_apply_control.R
\name{get_acts}
\alias{get_acts}
\title{get_acts}
\usage{
get_acts(actions)
}
\arguments{
\item{actions}{ACTION array from gmse_apply_ style output.}
}
\value{
A vector with six elements, being the sum of scaring, culling, castration, feeding, helping, and tend_crop actions, across all users.
}
\description{
Takes an ACTION array from gmse_apply_ style output, and returns a vector of the total number of actions summed across stakeholders.
}
\details{
Takes an ACTION array from gmse_apply_ style output, and returns a vector of the total number of actions summed across stakeholders.
}
\examples{
\dontrun{
step1 <- gmse_apply(get_res = "Full", scaring = TRUE, tend_crops = TRUE);
get_acts(step1$ACTION)

}
}
