% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmse_gui.R
\name{gmse_gui}
\alias{gmse_gui}
\title{GMSE GUI function}
\usage{
gmse_gui()
}
\value{
A browser should immediately open with the gmse graphical user
interface
}
\description{
The gmse_gui function will call a browser-based graphical user interface 
(GUI) for the gmse function. The GUI will run simulations for a limited range
of parameter values and present results as plots.
}
\examples{
\dontrun{
sim <- gmse_gui();
}
}
