% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmse_apply_control.R
\name{init_man_control}
\alias{init_man_control}
\title{init_man_control}
\usage{
init_man_control(K = 5)
}
\arguments{
\item{costs}{updated manager costs, as set by user in the preceding time step.}

\item{old_output}{existing gmse_apply_summary() output data to append to.}
}
\value{
A list with three elements: gmse_list which is each K gmse_apply output list, summary, which is the output summary as produced by gmse_apply_summary() and observed_suggested which is the output of observed_suggested() as run on the final gmse_apply_INTERIM() call.
}
\description{
init_man_control is a "set-up" for "manager control" gmse_apply() runs. It essentially wraps gmse_apply() and gmse_apply(INTERIM) calls; running K-1 gmse_apply() iterations followed by a single gmse_apply_INTERIM() call. The last resulting gmse_apply_ list struture is one ready for a gmse_apply_UROM() call, having finished with a call to the manager model.
}
\details{
init_man_control is a "set-up" for "manager control" gmse_apply() runs. It essentially wraps gmse_apply() and gmse_apply(INTERIM) calls; running K-1 gmse_apply() iterations followed by a single gmse_apply_INTERIM() call. The last resulting gmse_apply_ list struture is one ready for a gmse_apply_UROM() call, having finished with a call to the manager model.
}
\examples{
\dontrun{
step1 <- gmse_apply(get_res = "Full", scaring = TRUE);

}
}
