% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{initialise}
\alias{initialise}
\title{Simulation initialisation}
\usage{
initialise(model, resource_quantity, resource_types = 1, rows, cols, move)
}
\arguments{
\item{model}{The type of model being applied (Currently only individual-based
-- i.e., 'agent-based' -- models are allowed)}

\item{resource_quantity}{Number of resources being model}

\item{resource_types}{Number of different types of resources}

\item{rows}{Location of individual on LANDSCAPE (y-axis)}

\item{cols}{Location of individual on LANDSCAPE (x-axis)}

\item{move}{Parameter affecting individual movement}
}
\value{
the_resources List of data frames of resources being modelled
}
\description{
Function to initialise the landscape of the G-MSE model
}

