% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{make_costs}
\alias{make_costs}
\title{cOST initialisation}
\usage{
make_costs(AGENTS, RESOURCES, res_opts, lnd_opts, min_cost)
}
\arguments{
\item{AGENTS}{The agent array}

\item{RESOURCES}{The resource array}
}
\description{
Function to initialise the utilities of the G-MSE model
}
