% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{make_interaction_table}
\alias{make_interaction_table}
\title{Initialise array of resource and landscape-level interactions}
\usage{
make_interaction_table(resources, landscape)
}
\arguments{
\item{resources}{the resource array}

\item{landscape}{the landscape array}
}
\description{
Initialise array of resource and landscape-level interactions
}
