% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{make_utilities}
\alias{make_utilities}
\title{Utility initialisation}
\usage{
make_utilities(AGENTS, RESOURCES)
}
\arguments{
\item{AGENTS}{The agent array}

\item{RESOURCES}{The resource array}
}
\description{
Function to initialise the utilities of the G-MSE model
}
