% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{manager_user_budgets}
\alias{manager_user_budgets}
\title{Manager and user budgets}
\usage{
manager_user_budgets(
  agents,
  manager_budget = 1000,
  user_budget = 1000,
  usr_budget_rng = 0,
  budget_col = 17
)
}
\arguments{
\item{agents}{The agents array}

\item{manager_budget}{The budget of a manager}

\item{user_budget}{What is the budget of a user}

\item{usr_budget_rng}{Uniform range of users budgets}

\item{budget_col}{Column where the budget is located}
}
\value{
An updated agents data frame with correct budgets
}
\description{
Initialise manager and user budgets
}
