% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\name{observation}
\alias{observation}
\title{Observation model}
\usage{
observation(resource = NULL, landscape = NULL, paras = NULL,
  agent = NULL, type = 0, fix_mark = FALSE, model = "IBM")
}
\arguments{
\item{resource}{Resource data frame at the start of the time step}

\item{landscape}{Data frame at the start of the time step}

\item{paras}{Vector of parameter values to read into the model}

\item{agent}{Agent data frame at the start of the time step}

\item{model}{The type of model being applied (Currently only individual-based
-- i.e., 'agent-based' -- models are allowed)}

\item{types}{The types of data being observed}
}
\value{
Data frame observations at the end of the time step
}
\description{
A simulation of techniques (e.g., capture-mark-recapture) for estimating
population size and poperties.
}

