% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmse_apply_control.R
\name{observed_suggested}
\alias{observed_suggested}
\title{observed_suggested}
\usage{
observed_suggested(dat)
}
\arguments{
\item{dat}{gmse_apply(get_res="Full") style list.}
}
\value{
A list with three elements: observed, culling and scaring, which is the observed number of resources, and the cost suggested for culling and scaring actions in the given gmse_apply_ run.
}
\description{
Function to extract observed number of resources and costs for actions "proposed" by the manager model. Should work with any gmse_apply(get_res="Full") object, but intended for use following a run of gmse_apply_INTERIM() or gmse_apply_UROM().
}
\details{
Function to extract observed number of resources and costs for actions "proposed" by the manager model. Should work with any gmse_apply(get_res="Full") object, but intended for use following a run of gmse_apply_INTERIM() or gmse_apply_UROM().
}
\examples{
\dontrun{
sim_old <- gmse_apply(get_res = "Full", scaring = TRUE);
observed_suggested(sim_old);
}
}
