% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI5_data.R
\docType{data}
\name{rec.n}
\alias{rec.n}
\title{R data for recruitment used in SI4 vignette}
\format{
a set of \code{gmse_apply} output examples
}
\usage{
rec.n
}
\description{
This object contains SI4 vignette simulation output
}
\author{
Brad Duthie
}
