% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resource}
\alias{resource}
\title{Resource model}
\usage{
resource(resource = NULL, landscape = NULL, paras = NULL,
  move_res = TRUE, model = "IBM")
}
\arguments{
\item{resource}{Resource data frame at the start of the time step}

\item{landscape}{Data frame at the start of the time step}

\item{paras}{Vector of parameter values to read into the model}

\item{move_res}{Determining whether or not resources should move}

\item{model}{The type of model being applied (Currently only individual-based
-- i.e., 'agent-based' -- models are allowed)}
}
\value{
Data frame of Resource 2 at the end of the time step
}
\description{
A population model of resource (including population) dynamics for 
a single time step.
}
