% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI4_data.R
\docType{data}
\name{sim_init_basic}
\alias{sim_init_basic}
\title{R data needed for quick load basic sim in the S4 vignette}
\format{a set of \code{gmse_apply} output examples}
\usage{
sim_init_basic
}
\description{
This object contains SI4 vignette simulation output
}
\author{
Brad Duthie
}
