% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI4_data.R
\docType{data}
\name{sim_old_2_LAND}
\alias{sim_old_2_LAND}
\title{R data needed to show vignette example of LAND 2}
\format{a set of \code{gmse_apply} output examples}
\usage{
sim_old_2_LAND
}
\description{
This object contains simulation output for SI4 for LAND 1
}
\author{
Brad Duthie
}
