% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI4_data.R
\docType{data}
\name{sim_sum_1}
\alias{sim_sum_1}
\title{R data needed to show a summary of example simulation 1}
\format{a set of \code{gmse_apply} output examples}
\usage{
sim_sum_1
}
\description{
This object contains simulation the simulation 1 example in S4
}
\author{
Brad Duthie
}
