% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI4_data.R
\docType{data}
\name{sim_sum_2}
\alias{sim_sum_2}
\title{R data needed to show a summary of example simulation 2}
\format{a set of \code{gmse_apply} output examples}
\usage{
sim_sum_2
}
\description{
his object contains simulation the simulation 2 example in S4
}
\author{
Brad Duthie
}
