% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI5_data.R
\docType{data}
\name{ssb.n}
\alias{ssb.n}
\title{R data for spawning stock biomass used in SI4 vignette}
\format{a set of \code{gmse_apply} output examples}
\usage{
ssb.n
}
\description{
This object contains SI4 vignette simulation output
}
\author{
Brad Duthie
}
