% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\title{User model}
\usage{
user(resource = NULL, agent = NULL, landscape = NULL, paras = NULL,
  cost = NULL, action = NULL, model = "IBM")
}
\arguments{
\item{resource}{Resource data frame at the start of the time step}

\item{landscape}{Data frame at the start of the time step}

\item{paras}{Vector of parameter values to read into the model}

\item{model}{The type of model being applied (Currently only individual-based
-- i.e., 'agent-based' -- models are allowed)}
}
\value{
Data frames of user output at the end of the time step
}
\description{
A model of user decisions for a single time step
}
