% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{utility_layer}
\alias{utility_layer}
\title{Utility layer for initialisation}
\usage{
utility_layer(agent_IDs, agent_number, res_types)
}
\arguments{
\item{agent_IDs}{Vector of agent IDs to use (including -1 and -2)}

\item{agent_number}{The number of agents to use (length of agent_IDs)}

\item{res_types}{The number of unique resource types (cols 2-4 of RESOURCES)}
}
\description{
Function to initialise a layer of the UTILITY array of the G-MSE model
}
