% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{age_land}
\alias{age_land}
\title{Age landscape}
\usage{
age_land(landscape, landscape_ini, layer)
}
\arguments{
\item{landscape}{The name of the landscape being changed}

\item{landscape_ini}{The name of the original landscape replacing}

\item{layer}{The layer that is being affected on the landscape}

\item{model}{The type of model being applied (Currently only individual-based
-- i.e., 'agent-based' -- models are allowed)}
}
\description{
How does the landscape change on its own over generations?
For now, simply reverts a specified layer back to its original values
}
