% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anecdotal.R
\name{anecdotal}
\alias{anecdotal}
\title{Anecdotal model}
\usage{
anecdotal(resource = NULL, landscape = NULL, paras = NULL, agent = NULL,
  res_type = 1, samp_age = 1, agent_type = 0, type_cat = 1,
  move_agents = FALSE, model = "IBM")
}
\arguments{
\item{resource}{Resource data frame at the start of the time step}

\item{landscape}{Data frame at the start of the time step}

\item{paras}{Vector of parameter values to read into the model}

\item{agent}{Agent data frame at the start of the time step}

\item{samp_age}{Minimum age of the resource being sampled (default = 1)}

\item{agent_type}{The type of agent doing the observing (default = 0)}

\item{type_cat}{The category of agent type (first 4 columns) doing observing}

\item{move_agents}{Whether or not agents are moved during the run of anecodtal}

\item{model}{The type of model being applied (Currently only individual-based
-- i.e., 'agent-based' -- models are allowed)}

\item{types}{The types of data being observed}
}
\value{
Data frame observations at the end of the time step
}
\description{
A simulation of how many resources of a particular type are round each 
agent -- this produces a kind of anecdotal evidence for each agent around
their circle of view
}
