% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmse.R
\name{gmse}
\alias{gmse}
\title{General function for now}
\usage{
gmse(time_max = 100, land_dim_1 = 100, land_dim_2 = 100,
  res_movement = 1, remove_pr = 0, lambda = 0.05, agent_view = 10,
  agent_move = 50, res_birth_K = 10000, res_death_K = 400,
  edge_effect = 1, res_move_type = 2, res_birth_type = 2,
  res_death_type = 2, observe_type = 0, fixed_observe = 1,
  times_observe = 1, obs_move_type = 1, res_min_age = 1,
  res_move_obs = TRUE, Euclidean_dist = FALSE, plotting = TRUE,
  hunt = FALSE, start_hunting = 0, res_consume = 0.5, ga_popsize = 100,
  ga_mingen = 20, ga_seedrep = 10, ga_sampleK = 10, ga_chooseK = 2,
  ga_mutation = 0.1, ga_crossover = 0.1, move_agents = TRUE,
  max_ages = 5, minimum_cost = 10, user_budget = 1000,
  manager_budget = 1000, manage_target = 200, RESOURCE_ini = 100,
  scaring = FALSE, culling = TRUE, castration = FALSE, feeding = FALSE,
  help_offspring = FALSE, tend_crops = FALSE, kill_crops = FALSE,
  stakeholders = 4, manage_caution = 1, land_ownership = FALSE,
  manage_freq = 1, converge_crit = 100)
}
\description{
Function to run G-MSE model
}
