% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmse_apply.R
\name{gmse_apply}
\alias{gmse_apply}
\title{GMSE apply simulation}
\usage{
gmse_apply(time_max = 100)
}
\arguments{
\item{time_max}{This value sets the maximum number of time steps for a simulation. There are no constraints for length of time that a simulation can run. The default is 100 time steps.}
}
\description{
The gmse function is the the primary function to call to run a simulation.
It calls other functions that run resource, observation, management, and user
models in each time step. Hence while individual models can be used on their
own, gmse() is really all that is needed to run a simulation.
}
\examples{
sim <- gmse_apply(time_max = 100);
}
