% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{make_agents}
\alias{make_agents}
\title{Agent initialisation}
\usage{
make_agents(model = "IBM", agent_number = 2, by_type = FALSE,
  type_counts = c(1, 1), move = 0, vision = 20, rows = 100,
  cols = 100)
}
\arguments{
\item{model}{The type of model being applied (Currently only individual-based
-- i.e., 'agent-based' -- models are allowed)}

\item{agent_number}{Number of agents in the model}

\item{by_type}{Whether individuals are defined by type (FALSE) or row (TRUE)}

\item{type_counts}{Vector of how many agents of each type (element)}

\item{move}{A parameter that affects agent movement}

\item{vision}{A parameter affecting the area around its location it perceives}

\item{rows}{Location of agent on LANDSCAPE (y-axis)}

\item{cols}{Location of agent on LANDSCAPE (x-axis)}

\item{move}{Parameter affecting individual movement}
}
\value{
the_agents Initialised data frame of agents being modelled
}
\description{
Function to initialise the agents of the G-MSE model
}
