% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{make_landscape}
\alias{make_landscape}
\title{Landscape initialisation}
\usage{
make_landscape(model, rows, cols, cell_types)
}
\arguments{
\item{model}{The type of model being applied (Currently only individual-based
-- i.e., 'agent-based' -- models are allowed)}

\item{rows}{The dimension of the other side of the landscape (e.g., Longitude)}

\item{cols}{The dimension of one side of the landscape (e.g., Latitude)}

\item{cell_types}{Scalar or vector of all possible types of landscape cells}
}
\value{
the_land A cols by rows landscape with randomly distributed cell types
}
\description{
Function to initialise the landscape of the G-MSE model
}

