% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{make_resource}
\alias{make_resource}
\title{Resource initialisation}
\usage{
make_resource(model = "IBM", resource_quantity = 100, resource_types = 1,
  rows = 100, cols = 100, move = 1, rm_pr = 0, lambda = 0)
}
\arguments{
\item{model}{The type of model being applied (Currently only individual-based
-- i.e., 'agent-based' -- models are allowed)}

\item{resource_quantity}{Number of resources being model}

\item{resource_types}{Number of different types of resources}

\item{rows}{Location of resource on LANDSCAPE (y-axis)}

\item{cols}{Location of resource on LANDSCAPE (x-axis)}

\item{move}{Parameter affecting individual movement}
}
\value{
the_resources Initialised data frame of resources being modelled
}
\description{
Function to initialise the resources of the G-MSE model
}

